DEV AbstractValidatingBehavior added
